Further development of the profile page.
More specifically, I have begun creating a modal that is used to
create auditions and new ensembles. In this commit, creating an
audition still needs to be formatted properly and a submit button
needs to be created. Additionally, creating a new ensemble has not
been developed further than displaying the modal.